🩹 modify: delete files
