--- src/pages/googleapi.html
@@ -1,12 +1,0 @@
-<!doctype html>
-<html lang="ko-KR">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Document</title>
-    <script type="module" src="/src/typescript/googleapi.ts"></script>
-  </head>
-  <body>
-    <h1></h1>
-  </body>
-</html>
