+++ src/pages/main-page.html
@@ -4,8 +4,8 @@
     <meta charset="UTF-8" />
     <title>뉴스 카드 리스트</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script src="https://cdn.tailwindcss.com"></script>
-    <script src="/src/typescript/main-page-test.ts" type="module"></script>
+    <script type="module" src="/src/main.ts"></script>
+    <script src="/src/typescript/main-page.ts" type="module"></script>
   </head>
   <body class="bg-gray-100 flex flex-col min-h-screen">
     <header class="bg-white border-b w-full fixed top-0 left-0 z-50">
