+++ src/pages/search-page.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <title>뉴스 카드 리스트</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script src="https://cdn.tailwindcss.com"></script>
+    <script type="module" src="/src/main.ts"></script>
     <script src="/src/typescript/search-page.ts" type="module"></script>
   </head>
   <body class="bg-gray-100 p-6">
